att project
change API for mobile integration

+++ api/src/main/java/com/servicesystem/api/application/payload/response/ReviewsNoteResponse.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class ReviewsNoteResponse {
     
-    private Integer note;
+    private Double note;
     private UUID serviceProvidedId;
     private Integer numReviews;
 }
